move status related code to separate file
