+++ common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './Interfaces';
+export * from './StatusBar';
 export * from './Utils';
